Reflections in 2017
When this book, Violent Python was written, it may have been some incredible stuff. But as of 2017, a lot of the APIs that these programming projects are taking advantage of (Twitter and Google in Chapter 6), are no longer in use.

Furthermore, don't expect to be pulling off some crazy "Zero Day BS" on anyone with these scripts. Many of the author's work have been inspired by the actions of notorious convicted cybercriminals. That means, those vulnerabilities that made their names have been patched. And the Mitnick Attack? That was back in the days of America Online, Juno, and not everyone had a dial-up modem. 14.4 kbps was SCREAMING hot bandwidth back then. 

However, coding these timeless exploits ended up to be a lot of fun. If you wanted to understand the "Under the hood" behavior of a packet sniffer, a exif viewer, a bruteforcing tool, the C2 of a botnet all adapted into a Pythonic readable language, this book will give you a lot of inner understanding. 

Despite the syntax err